ig_agent_Cursor/requirements.txt
# Instagram AI Agent Dependencies
instagrapi==2.0.0
openai==1.3.0
anthropic==0.7.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncio==3.4.3
aiohttp==3.9.0
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.0
sqlalchemy==2.0.23
sqlite3
pandas==2.1.4
numpy==1.25.2
python-telegram-bot==20.7
flask==3.0.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
